Added parent class Contact for inheritance
